tests, refactor: removed OOBI resolution and added test coverage
We should not be using the OOBI parameter to resolve keystate to then allow for DID artifact resolution. Instead we should only be generating did.json and keri.cesr artifacts for KELs, TELs, and ACDCs from only the local keystore.

Building a did:webs DID artifact generator for any resolved KERI AID is a separate project and is out of scope for this resolver.

Many tests were added in this commit which caused many small refactorings to make things more testable and to make consistent APIs.