fix,tests: meta wrapping bug; did path bug; add logging; test dynamic gen ends
This adds integration tests for the endpoints that dynamically generate did.json and keri.cesr files. It also adds a feature for auto-wrapping a did.json document in did resolution metadata when the metadata is requested, which is useful and avoids unnecessary error responses when static did.json files do not have resolution metadata.

This also adds request logging for did.json and keri.cesr file resolution for log monitoring usability.