fix: use hab.replay for delegation chain; add missing ACDC anchor
The delegation chain is necessary when using delegated identifiers. This fixes that.

The missing ACDC anchor was a long-standing bug that prevented generation of both equivalentIds and alsoKnownAs based on the keri.cesr since the anchor tieing the designated aliases ACDC to the KEL for the AID being resolved was missing from the keri.cesr CESR stream.