fix: reset consumed request stream when debug logging
The prior version of the RequestLoggerMiddleware would consume the request body stream with req.stream.read() which broke things because the stream is a one-shot IO read meaning that the stream must be reset in order to be later consumed by the Falcon HTTP request handler. The logging was also updated to use placeholders rather than f-strings for performance.